+++ src/jaxnasium/_registry.py
@@ -4,12 +4,12 @@ from dataclasses import dataclass, field
 from typing import Any, Dict, List, Literal, Type
 
 from ._environment import Environment
-from ._wrappers import LogWrapper
-from ._wrappers_ext import (
+from .wrappers import (
     BraxWrapper,
     GymnaxWrapper,
     JaxMARLWrapper,
     JumanjiWrapper,
+    LogWrapper,
     NavixWrapper,
     PgxWrapper,
     Wrapper,
@@ -156,7 +156,7 @@ class Registry:
             elif package == "xminigrid":
                 import xminigrid
 
-                env, env_params = xminigrid.make(env_name, **env_kwargs)
+                env, env_params = xminigrid.make(env_name, **env_kwargs)  # type: ignore
                 if wrap:
                     return _wrap_env(env, xMinigridWrapper, _params=env_params)
                 return env  # type: ignore
